Make SetParents accommodate Match's limitations
The motivation here was an attempt to do a fact search (from a rule
condition) for a location's parents.  SetParents was giving a []string
as the property value, but Match can't match a []interface of strings
to a []string.  Instead of improving Match, we change SetParents to
provide a []interface of strings.