Start at support for Nanomsg
In the beginning, we supported only 0MQ.  More generally, we built
things to deal with messages instead of (say) HTTP requests.  So
supporting Nanomsg now is pretty easy.

We wrap Nanomsg around `service.Service` just as we wrapped HTTP
around it.

See 'service.TestNanomsg' for some examples.  That test includes a
Javascript function 'nanomsg()', which a rule action can use to
publish a nanomsg (so to speak).

We could build out a lot more gear for Nanomsg.  For example, we could
have an action interpreter which is designed to publish messages in a
conveniently.