Fix bug re "nil condition" due to typed nil; closes #37
When loc.ProcessEvent() was called from a rule action, the second
returned value is nil.  The type is *Condition (a concrete type) and
not an error (which is of course an interface).  Note that a
*Condition is in fact an error (since *Condition implements Error()).

See

  https://golang.org/doc/faq#nil_error

The code incorrectly checked to see if a *Condition was nil, which it
never is.  Instead, it can be (*Condition)(nil).

It's probably still reasonable for loc.ProcessEvent() to return the
concrete type *Condition because that's really what's intended.
However, maybe *Condition shouldn't implement error, which probably
deserves the most blame for this confusion.